refactor: layouts folder structure
