+++ apps/app/pages/[workspaceSlug]/settings/features.tsx
@@ -1,19 +1,21 @@
 import React from "react";
 
 import { useRouter } from "next/router";
+
 import useSWR from "swr";
 
 // lib
-import type { GetServerSideProps, NextPage } from "next";
 import { requiredWorkspaceAdmin } from "lib/auth";
-// constants
 // services
 import workspaceService from "services/workspace.service";
 // layouts
-import SettingsLayout from "layouts/settings-layout";
+import AppLayout from "layouts/app-layout";
 // ui
 import { Button } from "components/ui";
 import { BreadcrumbItem, Breadcrumbs } from "components/breadcrumbs";
+// types
+import type { GetServerSideProps, NextPage } from "next";
+// fetch-keys
 import { WORKSPACE_DETAILS } from "constants/fetch-keys";
 
 type TFeatureSettingsProps = {
@@ -35,11 +37,9 @@ const FeaturesSettings: NextPage<TFeatureSettingsProps> = (props) => {
 
   return (
     <>
-      <SettingsLayout
-        memberType={{
-          ...props,
-        }}
-        type="workspace"
+      <AppLayout
+        settingsLayout="workspace"
+        memberType={props}
         breadcrumbs={
           <Breadcrumbs>
             <BreadcrumbItem
@@ -141,7 +141,7 @@ const FeaturesSettings: NextPage<TFeatureSettingsProps> = (props) => {
             </div>
           </div>
         </section>
-      </SettingsLayout>
+      </AppLayout>
     </>
   );
 };
